Wire customer page and admin, not allow authenticated users to go to admin pages

+++ administrator/views.py
@@ -14,6 +14,9 @@ import datetime
 # Administrator function
 @login_required(login_url= '/accounts/login/')
 def administrator(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     total_users = User.objects.all().count()
     # Excursions queries
     all_excursion_orders = ExcursionOrder.objects.all()
@@ -57,6 +60,9 @@ def administrator(request):
 # Query all rentals for admin
 @login_required(login_url= '/accounts/login/')
 def admin_rentals(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     rentals = Rentals.objects.all()
     context = {'rentals': rentals}
     return render(request, 'administrator/rentals/admin_rentals.html', context)
@@ -65,6 +71,9 @@ def admin_rentals(request):
 
 @login_required(login_url= '/accounts/login/')
 def add_rentals(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     form = RentalForm()
     if request.method == 'POST':
         user = request.user
@@ -81,6 +90,9 @@ def add_rentals(request):
 
 @login_required(login_url= '/accounts/login/')
 def edit_rentals(request, pk):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     rentals = Rentals.objects.get(id=pk)
     form = RentalForm(instance=rentals)
     formPhotos = RentalFormPhotos()
@@ -107,6 +119,9 @@ def edit_rentals(request, pk):
 
 @login_required(login_url= '/accounts/login/')
 def delete_rentals_photos(request, pk):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     photo = Rental_photos.objects.get(id=pk)
     if request.method == 'POST':
         photo.delete()
@@ -119,6 +134,9 @@ def delete_rentals_photos(request, pk):
 
 @login_required(login_url= '/accounts/login/')
 def delete_rentals(request, pk):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     rental = Rentals.objects.get(id=pk)
     title = rental.title
     if request.method == 'POST':
@@ -134,6 +152,9 @@ def delete_rentals(request, pk):
 # Query all excursions for admin
 @login_required(login_url= '/accounts/login/')
 def admin_excursions(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     excursions = Excursions.objects.all()
     context = {'excursions': excursions}
     return render(request, 'administrator/excursions/admin_excursions.html', context)
@@ -141,6 +162,9 @@ def admin_excursions(request):
 # Add excursion
 @login_required(login_url= '/accounts/login/')
 def add_excursions(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     form = ExcursionForm()
     if request.method == 'POST':
         user = request.user
@@ -157,6 +181,9 @@ def add_excursions(request):
 # Edit excursion and add more photos
 @login_required(login_url= '/accounts/login/')
 def edit_excursions(request, pk):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     excursions = Excursions.objects.get(id=pk)
     form = ExcursionForm(instance=excursions)
     formPhotos = ExcursionFormPhotos()
@@ -184,6 +211,9 @@ def edit_excursions(request, pk):
 
 @login_required(login_url= '/accounts/login/')
 def delete_excursions(request, pk):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     excursion = Excursions.objects.get(id=pk)
     title = excursion.title
     if request.method == 'POST':
@@ -197,6 +227,9 @@ def delete_excursions(request, pk):
 
 @login_required(login_url= '/accounts/login/')
 def delete_excursions_photos(request, pk):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have persmision to access that page')
+        return redirect('home')
     photo = Photos.objects.get(id=pk)
     if request.method == 'POST':
         photo.delete()
